proto: build protobuf bindings with Prost/Tonic
This patch adds submodule dependencies on `tensorboard` for the main
protos and `tensorboard-data-server` for the gRPC service, along with a
build step to compile the interface definitions to Rust bindings. In the
future, we may want to make local changes to these protos, at which
point we can consider changing the submodule strategy.

Test Plan:
Running `cargo doc` and viewing the documentation shows that a new
module `rustboard_server::proto::tensorboard` contains the desired
symbols (like `Summary`), and likewise for the `data` submodule.