logdir: load runs in parallel with scoped threads
This patch resolves a TODO to add run-level parallelism to the logdir
loader, courtesy of `crossbeam` scoped threads.

Benchmark numbers compared to stock TensorBoard (measured by adding a
`std::process::exit(77)` to the end of `LogdirLoader::reload`):

  - 96 times faster on `mnist` (37 ms ± 4 ms)
  - 100 times faster on `nndiv_100m` (3 minutes 41.5 seconds)
  - 240 times faster on `edge_cgan` (4.64 s ± 0.29 s)

The `mnist` dataset is of course a microbenchmark. The other two are
just about factor-2 slower than the original `rustboard`, which is not
too shabby given that (a) this incarnation parses the protos with
`prost` rather than a hand-rolled parser, (b) this incarnation properly
handles truncation and growing files, and (c) this incarnation is
concurrent and live-updating, with an actual gRPC server.