server: implement `list_tensors`, `read_tensors`
Forked from the scalars code, with search-and-replace plus a handful of
lines changed (e.g., need to `.clone()` the proto in `read_tensors`
since `f64` is `Copy` but `TensorProto` isn’t). Can pull out a generic
implementation later, as in the Python code; let’s see how the blob code
informs things.

Test Plan:
Tested with the histograms and text plugins in TensorBoard; seems to
work.